+++ inc/class-main.php
@@ -140,8 +140,7 @@ class Main {
 			self::$assets_version = THEMEISLE_BLOCKS_VERSION;
 		}
 
-		add_action( 'init', array( $this, 'autoload_classes' ), 11 );
-		add_action( 'init', array( $this, 'register_blocks' ), 11 );
+		add_action( 'init', array( $this, 'autoload_classes' ), 9 );
 		add_action( 'enqueue_block_editor_assets', array( $this, 'enqueue_block_editor_assets' ) ); // Don't change the priority or else Blocks CSS will stop working.
 		add_action( 'enqueue_block_assets', array( $this, 'enqueue_block_frontend_assets' ) );
 		add_filter( 'script_loader_tag', array( $this, 'filter_script_loader_tag' ), 10, 2 );
@@ -155,11 +154,6 @@ class Main {
 				}
 			}
 		);
-		if ( version_compare( floatval( get_bloginfo( 'version' ) ), '5.8', '>=' ) ) {
-			add_filter( 'block_categories_all', array( $this, 'block_categories' ) );
-		} else {
-			add_filter( 'block_categories', array( $this, 'block_categories' ) );
-		}
 	}
 
 	/**
@@ -169,6 +163,12 @@ class Main {
 	 * @access  public
 	 */
 	public function autoload_classes() {
+		load_plugin_textdomain( 'otter-blocks', false, basename( OTTER_BLOCKS_PATH ) . '/languages' );
+
+		if ( class_exists( '\ThemeIsle\GutenbergBlocks\Registration' ) ) {
+			\ThemeIsle\GutenbergBlocks\Registration::instance();
+		}
+
 		$classnames = array(
 			'\ThemeIsle\GutenbergBlocks\CSS\Block_Frontend',
 			'\ThemeIsle\GutenbergBlocks\CSS\CSS_Handler',
@@ -195,8 +195,6 @@ class Main {
 			}
 		}
 
-		load_plugin_textdomain( 'otter-blocks', false, basename( OTTER_BLOCKS_PATH ) . '/languages' );
-
 		if ( class_exists( '\ThemeIsle\GutenbergBlocks\Blocks_CSS' ) && get_option( 'themeisle_blocks_settings_css_module', true ) ) {
 			\ThemeIsle\GutenbergBlocks\Blocks_CSS::instance();
 		}
@@ -207,162 +205,6 @@ class Main {
 
 		if ( class_exists( '\ThemeIsle\GutenbergBlocks\Blocks_Export_Import' ) ) {
 			\ThemeIsle\GutenbergBlocks\Blocks_Export_Import::instance();
-		}
-	}
-
-	/**
-	 * Autoload server side blocks.
-	 *
-	 * @since   1.0.0
-	 * @access  public
-	 */
-	public function register_blocks() {
-		// Handle Front:
-		// Circle Counter
-		// Countdown
-		// Form
-		// Google Map
-		// Leaflet Map - Also L is undefined error.
-		// Lottie
-		// Popup
-		// Progress Bar
-		// Section CSS file causing debugging errors.
-		$dynamic_blocks = array(
-			'about-author'              => '\ThemeIsle\GutenbergBlocks\Render\About_Author_Block',
-			'add-to-cart-button'        => '\ThemeIsle\GutenbergBlocks\Render\Add_To_Cart_Button_Block',
-			'form-nonce'                => '\ThemeIsle\GutenbergBlocks\Render\Form_Nonce_Block',
-			'google-map'                => '\ThemeIsle\GutenbergBlocks\Render\Google_Map_Block',
-			'leaflet-map'               => '\ThemeIsle\GutenbergBlocks\Render\Leaflet_Map_Block',
-			'plugin-card'               => '\ThemeIsle\GutenbergBlocks\Render\Plugin_Card_Block',
-			'posts-grid'                => '\ThemeIsle\GutenbergBlocks\Render\Posts_Grid_Block',
-			'review'                    => '\ThemeIsle\GutenbergBlocks\Render\Review_Block',
-			'review-comparison'         => '\ThemeIsle\GutenbergBlocks\Render\Review_Comparison_Block',
-			'sharing-icons'             => '\ThemeIsle\GutenbergBlocks\Render\Sharing_Icons_Block',
-			'woo-comparison'            => '\ThemeIsle\GutenbergBlocks\Render\Woo_Comparison_Block',
-			'product-add-to-cart'       => '\ThemeIsle\GutenbergBlocks\Render\Product_Add_To_Cart_Block',
-			'product-images'            => '\ThemeIsle\GutenbergBlocks\Render\Product_Images_Block',
-			'product-meta'              => '\ThemeIsle\GutenbergBlocks\Render\Product_Meta_Block',
-			'product-price'             => '\ThemeIsle\GutenbergBlocks\Render\Product_Price_Block',
-			'product-rating'            => '\ThemeIsle\GutenbergBlocks\Render\Product_Rating_Block',
-			'product-related-products'  => '\ThemeIsle\GutenbergBlocks\Render\Product_Related_Products_Block',
-			'product-short-description' => '\ThemeIsle\GutenbergBlocks\Render\Product_Short_Description_Block',
-			'product-stock'             => '\ThemeIsle\GutenbergBlocks\Render\Product_Stock_Block',
-			'product-tabs'              => '\ThemeIsle\GutenbergBlocks\Render\Product_Tabs_Block',
-			'product-title'             => '\ThemeIsle\GutenbergBlocks\Render\Product_Title_Block',
-			'product-upsells'           => '\ThemeIsle\GutenbergBlocks\Render\Product_Upsells_Block',
-		);
-
-		$blocks = array(
-			'about-author',
-			'accordion',
-			'accordion-item',
-			'add-to-cart-button',
-			'advanced-column',
-			'advanced-columns',
-			'advanced-heading',
-			'business-hours',
-			'business-hours-item',
-			'button',
-			'button-group',
-			'circle-counter',
-			'countdown',
-			'flip',
-			'font-awesome-icons',
-			'form',
-			'form-input',
-			'form-nonce',
-			'form-textarea',
-			'google-map',
-			'icon-list',
-			'icon-list-item',
-			'leaflet-map',
-			'lottie',
-			'plugin-card',
-			'popup',
-			'posts-grid',
-			'pricing',
-			'product-add-to-cart',
-			'product-images',
-			'product-meta',
-			'product-price',
-			'product-rating',
-			'product-related-products',
-			'product-short-description',
-			'product-stock',
-			'product-tabs',
-			'product-title',
-			'product-upsells',
-			'progress-bar',
-			'review',
-			'review-comparison',
-			'service',
-			'sharing-icons',
-			'slider',
-			'tabs',
-			'tabs-item',
-			'testimonials',
-			'woo-comparison',
-		);
-
-		foreach ( $blocks as $block ) {
-			$block_path    = OTTER_BLOCKS_PATH . '/build/blocks/' . $block;
-			$metadata_file = trailingslashit( $block_path ) . 'block.json';
-			$style         = trailingslashit( $block_path ) . 'style.css';
-			$editor_style  = trailingslashit( $block_path ) . 'editor.css';
-
-			if ( ! file_exists( $metadata_file ) ) {
-				continue;
-			}
-
-			$metadata = [];
-
-			if ( function_exists( 'wpcom_vip_file_get_contents' ) ) {
-				$metadata = json_decode( wpcom_vip_file_get_contents( $metadata_file ), true );
-			} else {
-				$metadata = json_decode( file_get_contents( $metadata_file ), true ); // phpcs:ignore WordPressVIPMinimum.Performance.FetchingRemoteData.FileGetContentsUnknown
-			}
-
-			$asset_file = include OTTER_BLOCKS_PATH . '/build/blocks/blocks.asset.php';
-
-			if ( file_exists( $editor_style ) && ! empty( $metadata['editorStyle'] ) ) {
-				wp_register_style( 
-					$metadata['editorStyle'],
-					plugin_dir_url( $this->get_dir() ) . 'build/blocks/' . $block . '/editor.css',
-					[],
-					$asset_file['version']
-				);
-			}
-
-			if ( ! is_admin() && file_exists( $style ) && ! empty( $metadata['style'] ) ) {
-				wp_register_style( 
-					$metadata['style'],
-					plugin_dir_url( $this->get_dir() ) . 'build/blocks/' . $block . '/style.css',
-					[],
-					$asset_file['version']
-				);
-			}
-
-			if ( ! is_array( $metadata ) || empty( $metadata['name'] ) ) {
-				continue;
-			}
-
-			if ( isset( $dynamic_blocks[ $block ] ) ) {
-				$classname = $dynamic_blocks[ $block ];
-				$renderer  = new $classname();
-		
-				if ( method_exists( $renderer, 'render' ) ) {
-					register_block_type_from_metadata(
-						$metadata_file,
-						array(
-							'render_callback' => array( $renderer, 'render' ),
-						) 
-					);
-
-					continue;
-				}
-			}
-
-			register_block_type_from_metadata( $metadata_file );
 		}
 	}
 
@@ -1067,32 +909,6 @@ class Main {
 		}
 
 		return current( $defaults );
-	}
-
-	/**
-	 * Register our custom block category.
-	 *
-	 * @param array $categories All categories.
-	 *
-	 * @return mixed
-	 * @since   1.0.0
-	 * @access public
-	 * @link   https://wordpress.org/gutenberg/handbook/extensibility/extending-blocks/#managing-block-categories
-	 */
-	public function block_categories( $categories ) {
-		return array_merge(
-			$categories,
-			array(
-				array(
-					'slug'  => 'themeisle-blocks',
-					'title' => $this->name,
-				),
-				array(
-					'slug'  => 'themeisle-woocommerce-blocks',
-					'title' => __( 'WooCommerce Builder by Otter', 'otter-blocks' ),
-				),
-			)
-		);
 	}
 
 	/**
